chore(sanity): rename getPairListener => createPairListener
