+++ packages/sanity/src/core/store/_legacy/document/index.ts
@@ -1,7 +1,7 @@
 export * from './buffered-doc'
+export * from './createPairListener'
 export * from './document-pair'
 export * from './document-store'
-export * from './getPairListener'
 export * from './hooks'
 export * from './initialValue'
 export * from './listenQuery'
